[ADD] often used calculet main page
