+++ client/src/Router.js
@@ -5,7 +5,6 @@ import Register from "./pages/Register";
 import BookmarkBar from "./components/global-component/BookmarkBar";
 import Login from "./pages/Login";
 import SignUp from "./pages/SignUp";
-import OftenCalculet from "./pages/OftenCalculet";
 
 const AppRouter = () => {
   return (
@@ -17,8 +16,6 @@ const AppRouter = () => {
         <Route path="/register" element={<Register />} />
         <Route path="/login" element={<Login />} />
         <Route path="/sign-up" element={<SignUp />} />
-        {/* <Route path="/" element={<Calculet />}></Route> */}
-        <Route path="/" element={<OftenCalculet />}></Route>
       </Routes>
       <BookmarkBar />
     </BrowserRouter>
