+++ client/src/components/atom-components/Bookmark.js
@@ -25,26 +25,10 @@ const StyledLink = styled(Link)`
  *
  */
 function Bookmark({ info }) {
-  function updateCalculetCount(calculetId) {
-    if (calculetId !== Number(localStorage.getItem("previewCalculet"))) {
-      console.log("update!", calculetId);
-      localStorage.setItem("previewCalculet", calculetId);
-      localStorage.setItem("continueCnt", 1);
-    } else {
-      console.log("continue");
-      const cnt = localStorage.getItem("continueCnt");
-      localStorage.setItem("continueCnt", Number(cnt) + 1);
-    }
-  }
-
   return (
     <Positioner>
       {info.map((conts, index) => (
-        <StyledLink
-          to={"/" + conts.id}
-          key={index}
-          onClick={() => updateCalculetCount(conts.id)}
-        >
+        <StyledLink to={"/" + conts.id} key={index}>
           {conts.title}
         </StyledLink>
       ))}
