+++ client/src/pages/Calculet.js
@@ -11,10 +11,6 @@ import TextWhite from "../components/atom-components/TextWhite";
 import Recommend from "../components/global-component/Recommend";
 import CalculetHeader from "../components/calculet-block/CalculetHeader";
 
-// (임시) html 파일 string으로 읽어오기 위해 사용
-// -> 백엔드 연결 이후에는 http request로 계산기 정보들과 함께 받음
-// eslint-disable-next-line
-import srcCode from "raw-loader!../calculets/arithmetic-operation/arithmeticOperation.html";
 import {
   ContentLayout,
   FlexColumnLayout,
@@ -23,6 +19,7 @@ import {
 } from "../components/Layout";
 
 import axios from "axios";
+import updateCalculetCount from "../utils/UpdateCalculetCount";
 
 // 계산기 블록 배경
 const Positioner = styled.div`
@@ -101,7 +98,31 @@ function Calculet() {
 
   // id 없다면 초기화 (메인 페이지)
   if (id === undefined) {
-    id = 1;
+    // 자주 쓰는 계산기의 선정 기준
+    const STANDARD_CNT = 3;
+
+    // 연속 횟수
+    const continueCnt = localStorage.getItem("continueCnt");
+
+    // 이전 계산기
+    const previewCalculet = localStorage.getItem("previewCalculet");
+
+    // 만약 이전 계산기의 연속 횟수가 기준에 도달했다면 자주 쓰는 계산기 값 변경
+    if (Number(continueCnt) === STANDARD_CNT) {
+      localStorage.setItem("oftenCalculet", previewCalculet);
+    }
+
+    // 만약 자주 쓰는 계산기가 비어있다면 초기화
+    if (localStorage.getItem("oftenCalculet") === null) {
+      localStorage.setItem("oftenCalculet", 1);
+      localStorage.setItem("previewCalculet", 1);
+      localStorage.setItem("continueCnt", 1);
+    }
+
+    // 자주 쓰는 계산기 가져오기
+    id = localStorage.getItem("oftenCalculet");
+
+    console.log(id);
   }
 
   /**
@@ -133,6 +154,7 @@ function Calculet() {
    */
   useEffect(() => {
     loadCalculetObj();
+    updateCalculetCount(id);
   }, [id]);
 
   return (
